updates only

+++ app/Models/Juego.php
@@ -70,12 +70,12 @@ class Juego extends Base
         }
 
         $validator = Validator::make($request->all(), [
-            'nombre' => 'required|min:2|max:255' . $exp,
-            'descripcion' => 'required|min:10|max:255',
-            'desarrolladora' => 'required|min:2|max:255',
-            'fecha' => 'required|date_format:Y-m-d',
-            'generos' => 'required|array|between:1,5',
-            'generos.*' => 'required|distinct|exists:generos,slug',
+            'nombre' => 'nullable|min:2|max:255' . $exp,
+            'descripcion' => 'nullable|min:10|max:255',
+            'desarrolladora' => 'nullable|min:2|max:255',
+            'fecha' => 'nullable|date_format:Y-m-d',
+            'generos' => 'nullable|array|between:1,5',
+            'generos.*' => 'nullable|distinct|exists:generos,slug',
             'imagen' => 'required|mimes:jpg,jpeg,png|max:1024|',
         ]);
 
@@ -92,12 +92,12 @@ class Juego extends Base
         }
 
         $validator = Validator::make($request->all(), [
-            'nombre' => 'required|min:2|max:255' . $exp,
-            'descripcion' => 'required|min:10|max:255',
-            'desarrolladora' => 'required|min:2|max:255',
-            'fecha' => 'required|date_format:Y-m-d',
-            'generos' => 'required|array|between:1,5',
-            'generos.*' => 'required|distinct|exists:generos,slug',
+            'nombre' => 'nullable|min:2|max:255' . $exp,
+            'descripcion' => 'nullable|min:10|max:255',
+            'desarrolladora' => 'nullable|min:2|max:255',
+            'fecha' => 'nullable|date_format:Y-m-d',
+            'generos' => 'nullable|array|between:1,5',
+            'generos.*' => 'nullable|distinct|exists:generos,slug',
         ]);
 
         return $validator;
@@ -141,25 +141,42 @@ class Juego extends Base
         if ($id_juego == null) {
             return response()->json(['error' => 'Juego no encontrado']);
         } else {
+
+          if($request->input('nombre') != null) {
             $slug = $this->sluggable($request->nombre);
             $request->request->add(['slug' => $slug]);
+          }
 
+          if($request->input('desarrolladora') != null) {
             $desarrolladora = new Desarrolladora();
             $desarrolladora = $desarrolladora->similar_name($request->input('desarrolladora'));
             $request->merge(['desarrolladora' => $desarrolladora->id]);
+          }
 
-            $id_juego->fill($request->all())->save();
+          $id_juego->update($request->only([
+              'nombre',
+              'descripcion',
+              'desarrolladora',
+              'fecha',
+              'generos',
+              'slug'
+          ]));
+
+          if($request->input('desarrolladora') != null) {
             $id_juego->desarrolladoras()->update(['juegable_id' => $desarrolladora->id]);
+          }
 
+          if($request->input('generos') != null) {
             $class_genero = new Genero();
             $id_juego->generos()->detach();
             foreach($request->input('generos') as $genero) {
               $class_genero->findBySlug($genero);
               $id_juego->generos()->syncWithoutDetaching($class_genero->findBySlug($genero)->id);
             }
+          }
 
             $class_imagen = new Imagen();
-            $class_imagen->updati($id_juego['id'], $slug, $request->imagen, 'juegos');
+            $class_imagen->updati($id_juego['id'], $request->input('slug'), $request->imagen, 'juegos');
             return $id_juego;
         }
     }
@@ -170,29 +187,48 @@ class Juego extends Base
         if ($id_juego == null) {
             return response()->json(['error' => 'Juego no encontrado']);
         } else {
-            $slug = $this->sluggable($request->nombre);
-            $slug_antiguo = $request->input('slug');
 
-            $desarrolladora = new Desarrolladora();
-            $desarrolladora = $desarrolladora->similar_name($request->input('desarrolladora'));
-            $request->merge(['desarrolladora' => $desarrolladora->id]);
+            if($request->input('nombre') != null) {
+              $slug = $this->sluggable($request->nombre);
+              $slug_antiguo = $request->input('slug');
+              $request->request->add(['slug' => $slug]);
+              //Cambiar el nombre del archivo
+              $id = $this->where('slug', $request->input('slug'))->first()->id;
+            }
 
-            $request->request->add(['slug' => $slug]);
-            $id_juego->fill($request->all())->save();
-            $id_juego->desarrolladoras()->update(['juegable_id' => $desarrolladora->id]);
-            
-            //Cambiar el nombre del archivo
-            $id = $this->where('slug', $request->input('slug'))->first()->id;
+            if($request->input('desarrolladora') != null) {
+              $desarrolladora = new Desarrolladora();
+              $desarrolladora = $desarrolladora->similar_name($request->input('desarrolladora'));
+              $request->merge(['desarrolladora' => $desarrolladora->id]);
+            }
 
-            $class_genero = new Genero();
-            $id_juego->generos()->detach();
-            foreach($request->input('generos') as $genero) {
-              $class_genero->findBySlug($genero);
-              $id_juego->generos()->syncWithoutDetaching($class_genero->findBySlug($genero)->id);
+            $id_juego->update($request->only([
+                'nombre',
+                'descripcion',
+                'desarrolladora',
+                'fecha',
+                'generos',
+                'slug'
+            ]));
+
+            if($request->input('desarrolladora') != null) {
+              $id_juego->desarrolladoras()->update(['juegable_id' => $desarrolladora->id]);
+            }
+
+            if($request->input('generos') != null) {
+              $class_genero = new Genero();
+              $id_juego->generos()->detach();
+              foreach($request->input('generos') as $genero) {
+                $class_genero->findBySlug($genero);
+                $id_juego->generos()->syncWithoutDetaching($class_genero->findBySlug($genero)->id);
+              }
             }
 
             $class_imagen = new Imagen();
-            $class_imagen->rename($id, $slug_antiguo, $slug, 'juegos');
+
+            if($request->input('nombre') != null) {
+              $class_imagen->rename($id, $slug_antiguo, $request->input('slug'), 'juegos');
+            }
 
             return $id_juego;
         }
